+++ allennlp/commands/elmo.py
@@ -76,7 +76,8 @@ from overrides import overrides
 from allennlp.commands.subcommand import Subcommand
 from allennlp.common.checks import ConfigurationError
 from allennlp.common.tqdm import Tqdm
-from allennlp.common.util import lazy_groups_of, prepare_global_logging
+from allennlp.common.logging import prepare_global_logging
+from allennlp.common.util import lazy_groups_of
 from allennlp.data.token_indexers.elmo_indexer import ELMoTokenCharactersIndexer
 from allennlp.modules.elmo import _ElmoBiLm, batch_to_ids
 from allennlp.nn.util import remove_sentence_boundaries
